+++ ui/src/main/java/org/akraino/validation/ui/controller/ModelsViewsController.java
@@ -77,6 +77,102 @@ public class ModelsViewsController extends RestrictedBaseController {
         return new ModelAndView(defaultViewName);
     }
 
+    @RequestMapping(value = { "/registerlab" }, method = RequestMethod.GET)
+    public ModelAndView registerLab(HttpServletRequest request) {
+        final String defaultViewName = null;
+        return new ModelAndView(defaultViewName);
+    }
+
+    @RequestMapping(value = { "/unregisterlab" }, method = RequestMethod.GET)
+    public ModelAndView unRegisterLab(HttpServletRequest request) {
+        final String defaultViewName = null;
+        return new ModelAndView(defaultViewName);
+    }
+
+    @RequestMapping(value = { "/modifylab" }, method = RequestMethod.GET)
+    public ModelAndView modifyLab(HttpServletRequest request) {
+        final String defaultViewName = null;
+        return new ModelAndView(defaultViewName);
+    }
+
+    @RequestMapping(value = { "/getlabs" }, method = RequestMethod.GET)
+    public ModelAndView getLabs(HttpServletRequest request) {
+        final String defaultViewName = null;
+        return new ModelAndView(defaultViewName);
+    }
+
+    @RequestMapping(value = { "/registerblueprint" }, method = RequestMethod.GET)
+    public ModelAndView registerBlueprint(HttpServletRequest request) {
+        final String defaultViewName = null;
+        return new ModelAndView(defaultViewName);
+    }
+
+    @RequestMapping(value = { "/unregisterblueprint" }, method = RequestMethod.GET)
+    public ModelAndView unRegisterBlueprint(HttpServletRequest request) {
+        final String defaultViewName = null;
+        return new ModelAndView(defaultViewName);
+    }
+
+    @RequestMapping(value = { "/getblueprints" }, method = RequestMethod.GET)
+    public ModelAndView getBlueprints(HttpServletRequest request) {
+        final String defaultViewName = null;
+        return new ModelAndView(defaultViewName);
+    }
+
+    @RequestMapping(value = { "/registerlayer" }, method = RequestMethod.GET)
+    public ModelAndView registerLayer(HttpServletRequest request) {
+        final String defaultViewName = null;
+        return new ModelAndView(defaultViewName);
+    }
+
+    @RequestMapping(value = { "/unregisterlayer" }, method = RequestMethod.GET)
+    public ModelAndView unRegisterLayer(HttpServletRequest request) {
+        final String defaultViewName = null;
+        return new ModelAndView(defaultViewName);
+    }
+
+    @RequestMapping(value = { "/getlayers" }, method = RequestMethod.GET)
+    public ModelAndView getLayers(HttpServletRequest request) {
+        final String defaultViewName = null;
+        return new ModelAndView(defaultViewName);
+    }
+
+    @RequestMapping(value = { "/registertimeslot" }, method = RequestMethod.GET)
+    public ModelAndView registerTimeslot(HttpServletRequest request) {
+        final String defaultViewName = null;
+        return new ModelAndView(defaultViewName);
+    }
+
+    @RequestMapping(value = { "/gettimeslots" }, method = RequestMethod.GET)
+    public ModelAndView getTimeslots(HttpServletRequest request) {
+        final String defaultViewName = null;
+        return new ModelAndView(defaultViewName);
+    }
+
+    @RequestMapping(value = { "/registerblueprintinstance" }, method = RequestMethod.GET)
+    public ModelAndView registerBlueprintInstance(HttpServletRequest request) {
+        final String defaultViewName = null;
+        return new ModelAndView(defaultViewName);
+    }
+
+    @RequestMapping(value = { "/unregisterblueprintinstance" }, method = RequestMethod.GET)
+    public ModelAndView unRegisterBlueprintInstance(HttpServletRequest request) {
+        final String defaultViewName = null;
+        return new ModelAndView(defaultViewName);
+    }
+
+    @RequestMapping(value = { "/getblueprintinstances" }, method = RequestMethod.GET)
+    public ModelAndView getBlueprintInstances(HttpServletRequest request) {
+        final String defaultViewName = null;
+        return new ModelAndView(defaultViewName);
+    }
+
+    @RequestMapping(value = { "/modifyblueprintinstance" }, method = RequestMethod.GET)
+    public ModelAndView modifyBlueprintInstance(HttpServletRequest request) {
+        final String defaultViewName = null;
+        return new ModelAndView(defaultViewName);
+    }
+
     @RequestMapping(value = { "/logout.htm" }, method = RequestMethod.GET)
     public ModelAndView login() {
         Map<String, Object> model = new HashMap<>();