+++ ui/src/main/java/org/akraino/validation/ui/entity/Timeslot.java
@@ -81,4 +81,5 @@ public class Timeslot implements Serializable {
     public LabInfo getLabInfo() {
         return labInfo;
     }
+
 }